Refactor project into modular package structure
