+++ README.md
@@ -1,2 +1,17 @@
 # FunGrid
-Basic Grid Environment implemented in python and pygame.
+
+Basic grid environment implemented in Python and pygame.
+
+## Running the game
+
+Install the dependencies listed in `requirements.txt` (if available) and then run:
+
+```bash
+python FunGrid.py
+```
+
+Alternatively, you can execute the package entry point directly:
+
+```bash
+python -m fun_grid.main
+```
