+++ fun_grid/__init__.py
@@ -1,0 +1,6 @@
+"""Core package for the FunGrid game environment."""
+
+from .game import Game
+from .main import main
+
+__all__ = ["Game", "main"]
